Move chunk and finalize routes under media path prefix
