+++ app/api/[owner]/[repo]/[branch]/media/[name]/[path]/chunk/route.ts
